requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.1.2
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
